+++ src/ts/components/audiotracklistbox.ts
@@ -1,22 +1,21 @@
 import {ListBox} from './listbox';
 import {ListSelectorConfig} from './listselector';
 import {UIInstanceManager} from '../uimanager';
-import {SubtitleSwitchHandler} from '../subtitleutils';
+import {AudioTrackSwitchHandler} from '../autiotrackutils';
 
 /**
  * A element that is similar to a select box where the user can select a subtitle
  */
-export class SubtitleListBox extends ListBox {
+export class AudioTrackListBox extends ListBox {
   constructor(config: ListSelectorConfig = {}) {
     super(config);
     this.config = this.mergeConfig(config, {
-      cssClass: 'ui-subtitle-listbox',
+      cssClass: 'ui-audiotrack-listbox',
     }, this.config);
   }
 
   configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
-
-    new SubtitleSwitchHandler(player, this);
+    new AudioTrackSwitchHandler(player, this);
   }
 }