+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [develop]
 
 ### Added
+- New `ListBox` UI element to display multiple selectable items
 - Icon for `Subtitles`
 - Icon for `AudioTracks`
 - Demo section within the UI variants in the `Playground`